ランデブー鍵派生を専用モジュールへ分離
